examples/crewai/pyproject.toml
[project]
name = "workbench-crewai-examples"
version = "0.1.0"
description = "Workbench CrewAI integration examples"
requires-python = ">=3.10"
dependencies = [
    "crewai>=0.1.0",
    "crewai-tools[mcp]>=0.1.0",
    "langchain-openai>=0.0.5",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0",
    "langchain-mcp-adapters>=0.1.11",
    "langchain-anthropic>=1.0.0",
]
